Added new items.json models + cleanup

+++ src/StatisticsAnalysisTool/Models/SimpleItemData.cs
@@ -3,7 +3,7 @@
     public class SimpleItemData
     {
         public string UniqueName { get; set; }
-        public int? ItemValue { get; set; }
+        public double? ItemValue { get; set; }
         public string ShopCategory { get; set; }
         public string ShopSubCategory1 { get; set; }
     }
